Adding configurability of frequency cap (min/max) files' path.
This way those devices that have a special path for
frequency caps can overlay it in the given device's
resources for Settings package.
It's defaulting to the general paths that are used
in most of the kernels.
It's usable for example in endeavoru device tree
where the max cap can only be set through
a different sys path.

Change-Id: Ic3113222580ebe8cc8e43d3dad7a15e041448c53

Conflicts:
	res/values/config.xml
	src/com/android/settings/carbon/BootReceiver.java