SlimPie fixes
Settings: Add user protection and pie/navbar toggle to SlimActions

Add drawables for it and enable it for navbar, pie, hwkeys
and GlobalMenu and whatever other shorcutfeature want to have it
Keep the user save and do not allow to disable all navigations to keep
the device usable. User will be notified

In Settings user get the choice to enable the other navigation method. Of course
only on navbar devices due that hw key devices always have a working navigation

and a small duplicate cleanup

PS:
- added Jubas suggestions

Change-Id: I3d13a13bf54304b0207b83b54df5c37337bcdaf5

Settings: SlimPie give user ability to reduce trigger heights if IME keyboard shows (2/2)

Add a user option to reduce left and right trigger of SlimPie if keyboard shows up
(especially usefull for swipe keyboard which most keyboards are nowadays). Enabled
by default user can disable it in settings

PS:
- fix category string

rebase

Change-Id: Ia5ae621a5e63b2efd10180cfbe1876b14a63c866

Settings: Respect and handle shorcut custom icon resources (1/2)

If a activity shorcut has an extra icon resource or drawable we need to handle
it. Fix this for all out shortcut features. Most is centralized so a small change

fix as well wrong sizes of custom images in QS custom tile

Change-Id: I53c5b4a2cb503c11dd3bb21e16a69eda0fb36b31